Add MessageContractList and rename namespaces
