+++ src/CSharp/EasyMicroservices.ServiceContracts/EasyMicroservices.ServiceContracts.csproj
@@ -5,7 +5,7 @@
     <Platforms>AnyCPU;x64;x86</Platforms>
     <Authors>EasyMicroservices</Authors>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>0.0.0.2</Version>
+    <Version>0.0.0.3</Version>
     <Description>Service base contracts to request or response</Description>
     <Copyright>EasyMicroservices@gmail.com</Copyright>
     <PackageTags>contract,message,messagecontract,service</PackageTags>
