+++ src/CSharp/EasyMicroservices.ServiceContracts/ErrorContract.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace ServiceContracts
+namespace EasyMicroservices.ServiceContracts
 {
     /// <summary>
     /// The contract of Error when the service response has error
