+++ src/CSharp/EasyMicroservices.ServiceContracts/MessageContractExtensions.cs
@@ -1,5 +1,4 @@
-﻿using ServiceContracts;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace EasyMicroservices.ServiceContracts
 {
